nature_downloader/requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
lxml==4.4.1
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
selenium==3.141.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
